# swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IDateDto": {
				"properties": {
					"date": {
						"type": "string"
					},
					"value": {
						"type": "boolean"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"date",
					"value",
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IDate": {
				"properties": {
					"date": {
						"type": "string"
					},
					"value": {
						"type": "boolean"
					}
				},
				"required": [
					"date",
					"value"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "access_token",
				"in": "query"
			}
		}
	},
	"info": {
		"title": "rest-api",
		"version": "1.0.0",
		"description": "REST API Template App",
		"license": {
			"name": "MIT"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/dates": {
			"get": {
				"operationId": "GetDates",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IDateDto"
									},
									"type": "array"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Dates"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "months",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/dates/save": {
			"post": {
				"operationId": "AddDate",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IDateDto"
									},
									"type": "array"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Dates"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/IDate"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
		"/api/dates/delete": {
			"post": {
				"operationId": "DeleteDate",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double"
								}
							}
						},
						"description": "Ok"
					}
				},
				"tags": [
					"Dates"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"type": "number",
									"format": "double"
								},
								"type": "array"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:5000/"
		}
	]
}